move AboutUs component to ui/ directory
