+++ frontend/src/widgets/about-us/index.ts
@@ -1,1 +1,1 @@
-export { AboutUs } from './AboutUs';
+export { AboutUs } from './ui/AboutUs';
